Effect size conversion

es is used to convert one effect size to other effect size measures. Currently supports Cohen's d, corelation r, r-squared, Cohen's f, odds ratio, log odds ratio, and area-under-curve (auc). Also available as a Shiny app here: http://escal.site

es(d = NULL, r = NULL, R2 = NULL, f = NULL, oddsratio = NULL,
logoddsratio = NULL, auc = NULL, fishersz = NULL,
decimal = 3, msg = TRUE)

Arguments

d
a numeric vector containing Cohen's d effect size(s)

r
a numeric vector containing correlation r effect size(s)

R2
a numeric vector containing r-squared effect size(s)

f
a numeric vector containing Cohen's f effect size(s)

oddsratio
a numeric vector containing odds ratio effect size(s)

logoddsratio
a numeric vector containing log odds ratio effect size(s)

auc
a numeric vector containing area-under-curve effect size(s)

fishersz
a numeric vector containing fisher's z effect size(s)

decimal
a numeric vector indicating decimal places of output

msg
a boolean indicating whether to show input effect size(s)

Value

A dataframe with converted effect sizes

Details

Formulae for conversion

f = d / 2
r = d / sqrt(d^2 + 4)
d = (2 * r) / sqrt(1 - r^2)
R2 = r^2
oddsratio = exp(d / (sqrt(3) / pi))
logoddsratio = d / (sqrt(3) / pi)
auc = pnorm(d / sqrt(2), 0, 1)
fisher's z = 0.5 * (log(1 + r) - log(1 - r))} \cr } \note{ All conversions assume equal-sized groups. Effect size conventions: \cr Cohen's d: 0.20 (small), 0.50 (medium), .80 (large) (Cohen, 1992)
correlation r: .10 (small), .30 (medium), .50 (large)
R-squared: R2: .02 (small), .13 (medium), .26 (large) es(d = 0.3) es(r = c(0.1, 0.3)) Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009). Introduction to meta-analysis. Chichester, West Sussex, UK: Wiley.

Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.), Hillsdale, NJ: Erlbaum.

Rosenthal, R. (1994). Parametric measures of effect size. In H. Cooper & L. V. Hedges (Eds.), The Handbook of Research Synthesis. New York, NY: Sage.

Ruscio, J. (2008). A probability-based measure of effect size: Robustness to base rates and other factors. Psychological Methods, 13(1), 19-30. doi:10.1037/1082-989x.13.1.19 Hause Lin