Specification curve analysis (SCA) for linear regression

sca_lm is used to run specification curve analysis using linear regression. Fits every possible model for your specified dependent variable, predictor variables, and covariates

sca_lm(data, dv, ivs, covariates = NULL)

Arguments

data
dataset

dv
dependent variable (outcome variable) (character)

ivs
independent variable(s) or predictor(s) (character)

covariates
covariates (character)

Value

A tibble/data.table with results

Details

Details to follow

References

References to follow

Examples

# model with 1 covariate
m1 <- sca_lm(data = mtcars, dv = "mpg", ivs = c("cyl", "carb"), covariates = c("vs"))
m1[, c("modelformula", "estimate", "p.value", "es.r")] # model, beta, p value, effect size

#>              modelformula estimate p.value  es.r
#>  1:       mpg ~ carb + vs   20.061   0.000 0.838
#>  2:       mpg ~ carb + vs   -0.954   0.126 0.281
#>  3:       mpg ~ carb + vs    6.199   0.003 0.511
#>  4: mpg ~ cyl + carb + vs   41.107   0.000 0.877
#>  5: mpg ~ cyl + carb + vs   -2.981   0.000 0.714
#>  6: mpg ~ cyl + carb + vs   -0.636   0.155 0.266
#>  7: mpg ~ cyl + carb + vs   -1.786   0.385 0.164
#>  8:        mpg ~ cyl + vs   39.625   0.000 0.867
#>  9:        mpg ~ cyl + vs   -3.091   0.000 0.717
#> 10:        mpg ~ cyl + vs   -0.939   0.638 0.088

# model without covariates
m2 <- sca_lm(data = mtcars, dv = "mpg", ivs = c("cyl", "gear"))
m2[, c("modelformula", "estimate", "p.value", "es.r")]

#>        modelformula estimate p.value  es.r
#> 1:        mpg ~ cyl   37.885   0.000 0.958
#> 2:        mpg ~ cyl   -2.876   0.000 0.852
#> 3: mpg ~ cyl + gear   34.659   0.000 0.793
#> 4: mpg ~ cyl + gear   -2.743   0.000 0.806
#> 5: mpg ~ cyl + gear    0.652   0.477 0.133
#> 6:       mpg ~ gear    5.623   0.262 0.204
#> 7:       mpg ~ gear    3.923   0.005 0.480